requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
langserve==0.0.30
langchain==0.0.325
python-dotenv==1.0.0
openai==0.28.1
anthropic==0.3.11
transformers==4.34.1
torch==2.1.0
pillow==10.0.1